requirements.txt
APScheduler==3.5.1
Babel==2.5.3
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
dateutils==0.6.6
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.8.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
passlib==1.7.1
psycopg2==2.7.4
pycparser==2.18
python-dateutil==2.7.3
pytz==2018.4
redis==2.10.6
requests==2.18.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.7
tzlocal==1.5.1
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
